backend/requirements.txt
flask==2.0.1
azure-identity==1.12.0
azure-storage-blob==12.16.0
azure-cosmos==4.3.1
azure-search-documents==11.4.0
azure-ai-formrecognizer==3.2.1
azure-monitor-opentelemetry==1.0.0
python-dotenv==1.0.0
gunicorn==20.1.0
openai==1.3.0
beautifulsoup4==4.12.2
requests==2.31.0
pypdf==3.17.1
